test: Move 'custom grouping by tag' documentation tests to separate folder
To avoid cluttering up the main Filter tests
