+++ tests/Query/Filter/TagsField.test.ts
@@ -1,18 +1,14 @@
-import { verifyAll } from 'approvals/lib/Providers/Jest/JestApprovals';
 import { GlobalFilter } from '../../../src/Config/GlobalFilter';
-import { FunctionField } from '../../../src/Query/Filter/FunctionField';
-import { groupHeadingsForTask } from '../../CustomMatchers/CustomMatchersForGrouping';
 import type { FilteringCase } from '../../TestingTools/FilterTestHelpers';
 import { shouldSupportFiltering } from '../../TestingTools/FilterTestHelpers';
 import { TagsField } from '../../../src/Query/Filter/TagsField';
 import { DescriptionField } from '../../../src/Query/Filter/DescriptionField';
-import { SampleTasks, fromLine } from '../../TestHelpers';
+import { fromLine } from '../../TestHelpers';
 import { Sort } from '../../../src/Query/Sort';
 import { TaskBuilder } from '../../TestingTools/TaskBuilder';
 import { expectTaskComparesAfter, expectTaskComparesBefore } from '../../CustomMatchers/CustomMatchersForSorting';
 import type { Grouper } from '../../../src/Query/Grouper';
 import { TaskGroups } from '../../../src/Query/TaskGroups';
-import { verifyMarkdownForDocs } from '../../TestingTools/VerifyMarkdownTable';
 
 describe('tag presence & absence', () => {
     it.each(['has tag', 'has tags'])('should have "%s" filtering', (filterLine: string) => {
@@ -692,60 +688,5 @@ describe('grouping by tag', () => {
         const tasks = taskLines.map((taskLine) => fromLine({ line: taskLine }));
 
         expect({ grouper, tasks }).groupHeadingsToBe(['(No tags)', '#tag1', '#tag1/tag3', '#tag2', '#tag2/tag4']);
-    });
-});
-
-describe('custom grouping by tag', () => {
-    const customGroups = [
-        [
-            'group by function task.tags',
-            'Like "group by tags" except that tasks with no tags have no heading instead of "(No tags)"',
-        ],
-        [
-            'group by function task.tags.join(", ")',
-            'Tasks with multiple tags are listed once, with a heading that combines all the tags. Separating with commas means the tags are clickable in the headings',
-        ],
-        [
-            'group by function task.tags.filter( (t) => t.includes("#context/"))',
-            'Only create headings for tags that contain "#context/"',
-        ],
-        [
-            'group by function task.tags.filter( (t) => ! t.includes("#tag"))',
-            'Create headings for all tags that do not contain "#tag"',
-        ],
-    ];
-
-    it('results', () => {
-        const tasks = SampleTasks.withRepresentativeTags();
-        verifyAll('Results of custom groupers', customGroups, (group) => {
-            const instruction = group[0];
-            const comment = group[1];
-            const grouper = new FunctionField().createGrouperFromLine(instruction);
-            const headings = groupHeadingsForTask(grouper!, tasks);
-            return `
-${instruction}
-${comment}
-=>
-${headings.join('\n')}
-====================================================================================
-`;
-        });
-    });
-
-    it('docs', () => {
-        let markdown = '';
-        for (const group of customGroups) {
-            const instruction = group[0];
-            const comment = group[1];
-            markdown += `
-~~~text
-${instruction}
-~~~
-
-- ${comment}.
-
-`;
-        }
-        verifyMarkdownForDocs(markdown);
     });
 });
